Sophisticate clean_job to remove jobs after active

+++ app/jobs/clean_job.rb
@@ -2,7 +2,26 @@ class CleanJob < ApplicationJob
   queue_as :default
 
   def perform(*args)
-    batch_range = 0..(Batch.active_marker-1)
-    Batch.where(marker: batch_range).destroy_all
+
+    start_batch = 1
+    end_batch = Batch.active_marker-1
+
+    purge_batch_range start_batch, end_batch
+
+    start_batch = Batch.active_marker+1
+    end_batch = Batch.current_marker-1
+
+    purge_batch_range start_batch, end_batch
+
   end
+
+  private
+
+    def purge_batch_range start_batch, end_batch
+      batch_range = start_batch..end_batch
+      return if batch_range.count.zero?
+
+      Batch.where(marker: batch_range).destroy_all
+    end
+
 end
